Move to `SmallStatBlock`
