+++ app/Filament/Server/Widgets/ServerOverview.php
@@ -2,10 +2,13 @@
 
 namespace App\Filament\Server\Widgets;
 
+use App\Enums\ContainerStatus;
+use App\Filament\Server\Components\SmallStatBlock;
 use App\Filament\Server\Components\StatBlock;
 use App\Models\Server;
 use Carbon\CarbonInterface;
 use Filament\Widgets\StatsOverviewWidget;
+use Illuminate\Support\Number;
 use Illuminate\Support\Str;
 
 class ServerOverview extends StatsOverviewWidget
@@ -27,6 +30,9 @@ class ServerOverview extends StatsOverviewWidget
                 ->extraAttributes([
                     'class' => 'overflow-x-auto',
                 ]),
+            SmallStatBlock::make('CPU', $this->cpuUsage()),
+            SmallStatBlock::make('Memory', $this->memoryUsage()),
+            SmallStatBlock::make('Disk', $this->diskUsage()),
         ];
     }
 
@@ -42,5 +48,50 @@ class ServerOverview extends StatsOverviewWidget
         $uptime = now()->subMillis($uptime)->diffForHumans(syntax: CarbonInterface::DIFF_ABSOLUTE, short: true, parts: 2);
 
         return "$status ($uptime)";
+    }
+
+    public function cpuUsage(): string
+    {
+        $status = ContainerStatus::tryFrom($this->server->retrieveStatus());
+
+        if ($status === ContainerStatus::Offline || $status === ContainerStatus::Missing) {
+            return 'Offline';
+        }
+
+        $data = collect(cache()->get("servers.{$this->server->id}.cpu_absolute"))->last(default: 0);
+        $cpu = Number::format($data, maxPrecision: 2, locale: auth()->user()->language) . ' %';
+
+        return $cpu . ($this->server->cpu > 0 ? ' / ' . Number::format($this->server->cpu, locale: auth()->user()->language) . ' %' : ' / ∞');
+    }
+
+    public function memoryUsage(): string
+    {
+        $status = ContainerStatus::tryFrom($this->server->retrieveStatus());
+
+        if ($status === ContainerStatus::Offline || $status === ContainerStatus::Missing) {
+            return 'Offline';
+        }
+
+        $latestMemoryUsed = collect(cache()->get("servers.{$this->server->id}.memory_bytes"))->last(default: 0);
+        $totalMemory = collect(cache()->get("servers.{$this->server->id}.memory_limit_bytes"))->last(default: 0);
+
+        $used = convert_bytes_to_readable($latestMemoryUsed);
+        $total = convert_bytes_to_readable($totalMemory);
+
+        return $used . ($this->server->memory > 0 ? ' / ' . $total : ' / ∞');
+    }
+
+    public function diskUsage(): string
+    {
+        $disk = collect(cache()->get("servers.{$this->server->id}.disk_bytes"))->last(default: 0);
+
+        if ($disk === 0) {
+            return 'Unavailable';
+        }
+
+        $used = convert_bytes_to_readable($disk);
+        $total = convert_bytes_to_readable($this->server->disk);
+
+        return $used . ($this->server->disk > 0 ? ' / ' . $total : ' / ∞');
     }
 }
