+++ resources/views/filament/components/server-small-data-block.blade.php
@@ -1,0 +1,9 @@
+<div
+    class="grid grid-flow-row w-full p-3 rounded-lg shadow-sm overflow-hidden ring-1 ring-gray-950/5 dark:bg-gray-900 dark:ring-white/10">
+    <span>
+        <span class="text-md font-medium text-gray-500 dark:text-gray-400">
+            {{ $getLabel() }}
+        </span>
+        <span class="text-md font-semibold">{{ $getValue() }}</span>
+    </span>
+</div>
